+++ src/imio/smartweb/common/ia/browser/static/webpack.config.js
@@ -1,0 +1,41 @@
+const path = require('path');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const CssMinimizerPlugin = require("css-minimizer-webpack-plugin");
+const TerserPlugin = require("terser-webpack-plugin");
+
+
+module.exports = {
+    mode: 'development',   
+    entry: {
+        ia_suggested_titles: './src/ia_suggested_titles.js',
+    },
+    output: {
+        filename: 'smartweb-common-[name]-compiled.js',
+        path: path.resolve(__dirname, ''),
+    },
+    module: {
+        rules: [
+            {
+                test: /\.less$/,
+                use: [
+                    MiniCssExtractPlugin.loader,
+                    'css-loader',
+                    'less-loader'
+                ],
+            },
+        ]
+    },
+    plugins: [
+        new MiniCssExtractPlugin({
+            filename: 'smartweb-common-[name]-compiled.css',
+        }),
+    ],
+    optimization: {
+        usedExports: true,
+        minimize: true,
+        minimizer: [
+          new CssMinimizerPlugin(),
+          new TerserPlugin()
+        ],
+      },
+};